[서버] GameServer 코드 일부 수정 및 출력창 오류 모두 정리 완료

+++ Server/Single Thread ASIO/source/Bomb.h
@@ -9,10 +9,10 @@ public:
 	Vector3D	pos_;
 	Vector3D	initialVelocity_;
 	int			index_;
-	double		rx_, ry_, rz_;
-	double		r_;
-	double		speed_;
-	double		yaw_, pitch_;
+	float		rx_, ry_, rz_;
+	float		r_;
+	float		speed_;
+	float		yaw_, pitch_;
 	BombType	bomb_type_;		// Type of bomb -> -1 : default, 0 : stun, 1 : explosion
 	
 public:
@@ -21,20 +21,20 @@ public:
 		bomb_type_ = NoBomb;
 		pos_.x = pos_.y=pos_.z = 0;
 		speed_ = 1600;
-		r_ = 0.5358935;
+		r_ = float(0.5358935);
 	}
 
 	Bomb(BombType type) :bomb_type_(type) {
 		in_use_ = true;
 		pos_.x = pos_.y = pos_.z = 0;
 		speed_ = 1600;
-		r_ = 0.5358935;
+		r_ = float(0.5358935);
 	}
 
 	Vector3D calculateInitialVelocity() {
-		double vx = speed_ * rx_;
-		double vy = speed_ * ry_;
-		double vz = speed_ * rz_;
+		float vx = speed_ * rx_;
+		float vy = speed_ * ry_;
+		float vz = speed_ * rz_;
 
 		return { vx, vy, vz };
 	}
